+++ packages/react-dom/src/__tests__/refs-test.js
@@ -369,24 +369,6 @@ describe('ref swapping', () => {
       });
     }).rejects.toThrow('Expected ref to be a function');
   });
-
-  // @gate !enableRefAsProp && www
-  it('undefined ref on manually inlined React element triggers error', async () => {
-    const container = document.createElement('div');
-    const root = ReactDOMClient.createRoot(container);
-    await expect(async () => {
-      await act(() => {
-        root.render({
-          $$typeof: Symbol.for('react.element'),
-          type: 'div',
-          props: {
-            ref: undefined,
-          },
-          key: null,
-        });
-      });
-    }).rejects.toThrow('Expected ref to be a function');
-  });
 });
 
 describe('root level refs', () => {