+++ packages/react-server-dom-esm/src/client/ReactFlightDOMClientBrowser.js
@@ -51,6 +51,7 @@ function createResponseFromOptions(options: void | Options) {
   return createResponse(
     options && options.moduleBaseURL ? options.moduleBaseURL : '',
     null,
+    null,
     options && options.callServer ? options.callServer : undefined,
     undefined, // encodeFormAction
     undefined, // nonce