+++ packages/react-server-dom-webpack/src/client/ReactFlightClientConfigBundlerNode.js
@@ -24,7 +24,7 @@ import {
 } from '../shared/ReactFlightImportMetadata';
 import {prepareDestinationWithChunks} from 'react-client/src/ReactFlightClientConfig';
 
-export type SSRModuleMap = {
+export type ServerConsumerModuleMap = {
   [clientId: string]: {
     [clientExportName: string]: ClientReference<any>,
   },
@@ -58,23 +58,23 @@ export function prepareDestinationForModule(
 }
 
 export function resolveClientReference<T>(
-  bundlerConfig: SSRModuleMap,
+  bundlerConfig: ServerConsumerModuleMap,
   metadata: ClientReferenceMetadata,
 ): ClientReference<T> {
   const moduleExports = bundlerConfig[metadata[ID]];
-  let resolvedModuleData = moduleExports[metadata[NAME]];
+  let resolvedModuleData = moduleExports && moduleExports[metadata[NAME]];
   let name;
   if (resolvedModuleData) {
     // The potentially aliased name.
     name = resolvedModuleData.name;
   } else {
     // If we don't have this specific name, we might have the full module.
-    resolvedModuleData = moduleExports['*'];
+    resolvedModuleData = moduleExports && moduleExports['*'];
     if (!resolvedModuleData) {
       throw new Error(
         'Could not find the module "' +
           metadata[ID] +
-          '" in the React SSR Manifest. ' +
+          '" in the React Server Consumer Manifest. ' +
           'This is probably a bug in the React Server Components bundler.',
       );
     }