+++ packages/shared/forks/ReactFeatureFlags.test-renderer.www.js
@@ -83,7 +83,6 @@ export const disableClientCache = true;
 export const enableServerComponentLogs = true;
 export const enableInfiniteRenderLoopDetection = false;
 
-export const enableRefAsProp = true;
 export const disableStringRefs = false;
 
 export const enableReactTestRendererWarning = false;